Chore #157990868 updating tests to fit recommendation
