+++ .gitignore
@@ -1,3 +1,5 @@
 __pycache__/
 env/
 *.pyc
+*.pytest_cache
+.pytest_cache
