+++ .pytest_cache/v/cache/lastfailed
@@ -4,6 +4,5 @@
   "tests/unit_test.py::ApiTest::test_api_can_get_request_by_id": true,
   "tests/unit_test.py::ApiTest::test_api_request_can_be_modified": true,
   "tests/unit_test.py::test_api_can_get_all_requests": true,
-  "tests/user_auth_test.py": true,
-  "tests/user_auth_test.py::UserTest::test_user_authentication": true
+  "tests/user_auth_test.py::TestingConfig::()": true
 }
