+++ .pytest_cache/v/cache/nodeids
@@ -1,1 +1,7 @@
-[]
+[
+  "tests/test_requests.py::ApiTest::test_api_can_create_request",
+  "tests/test_requests.py::ApiTest::test_api_can_get_all_requests",
+  "tests/test_requests.py::ApiTest::test_api_can_get_request_by_id",
+  "tests/test_requests.py::ApiTest::test_api_request_can_be_modified",
+  "tests/user_auth_test.py::UserTest::test_user_authentication"
+]
